+++ primefaces-integration-tests/src/test/java/org/primefaces/integrationtests/confirmdialog/ConfirmDialog001Test.java
@@ -34,6 +34,7 @@ import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.support.FindBy;
 import org.primefaces.selenium.AbstractPrimePage;
 import org.primefaces.selenium.AbstractPrimePageTest;
+import org.primefaces.selenium.PrimeExpectedConditions;
 import org.primefaces.selenium.PrimeSelenium;
 import org.primefaces.selenium.component.CommandButton;
 import org.primefaces.selenium.component.ConfirmDialog;
@@ -96,9 +97,12 @@ class ConfirmDialog001Test extends AbstractPrimePageTest {
         ConfirmDialog dialog = page.dialog;
         assertFalse(dialog.isVisible());
         page.confirm.click();
+        CommandButton noButton = dialog.getNoButton();
+        assertEquals("No", noButton.getText());
+        assertCss(noButton, "ui-button ui-widget ui-state-default ui-corner-all ui-button-text-only ui-confirmdialog-no ui-button-flat");
 
         // Act
-        dialog.getNoButton().click();
+        noButton.click();
 
         // Assert
         assertTrue(page.message.isEmpty());
@@ -113,9 +117,12 @@ class ConfirmDialog001Test extends AbstractPrimePageTest {
         ConfirmDialog dialog = page.dialog;
         page.confirm.click();
         assertTrue(dialog.isVisible());
+        CommandButton yesButton = dialog.getYesButton();
+        assertEquals("Yes", yesButton.getText());
+        assertCss(yesButton, "ui-button ui-widget ui-state-default ui-corner-all ui-button-text-only ui-confirmdialog-yes");
 
         // Act
-        PrimeSelenium.guardAjax(dialog.getYesButton()).click();
+        PrimeSelenium.guardAjax(yesButton).click();
 
         // Assert
         assertEquals("You have accepted", page.message.getMessage(0).getDetail());
@@ -130,6 +137,10 @@ class ConfirmDialog001Test extends AbstractPrimePageTest {
         ConfirmDialog dialog = page.dialog;
         assertFalse(dialog.isVisible());
         page.delete.click();
+        CommandButton noButton = dialog.getNoButton();
+        assertEquals("Keep this!", noButton.getText());
+        assertCss(noButton,
+                "ui-button ui-widget ui-state-default ui-corner-all ui-button-text-only ui-confirmdialog-no ui-button-flat bg-green-500 text-white");
 
         // Act
         dialog.getNoButton().click();
@@ -137,6 +148,9 @@ class ConfirmDialog001Test extends AbstractPrimePageTest {
         // Assert
         assertTrue(page.message.isEmpty());
         assertDialog(page, false);
+
+        // assert the buttons are back to normal
+        confirmNo(page);
     }
 
     @Test
@@ -147,13 +161,18 @@ class ConfirmDialog001Test extends AbstractPrimePageTest {
         ConfirmDialog dialog = page.dialog;
         assertFalse(dialog.isVisible());
         page.delete.click();
+        CommandButton yesButton = dialog.getYesButton();
+        assertEquals("Delete Me!", yesButton.getText());
+        assertCss(yesButton, "ui-button ui-widget ui-state-default ui-corner-all ui-button-text-only ui-confirmdialog-yes bg-red-500 text-white");
 
         // Act
-        PrimeSelenium.guardAjax(dialog.getYesButton()).click();
+        PrimeSelenium.guardAjax(yesButton).click();
 
         // Assert
         assertEquals("Record deleted", page.message.getMessage(0).getDetail());
         assertDialog(page, false);
+        // assert the buttons are back to normal
+        confirmYes(page);
     }
 
     @Test
@@ -210,6 +229,7 @@ class ConfirmDialog001Test extends AbstractPrimePageTest {
         else {
             assertClickableOrLoading(page.confirm);
             assertClickableOrLoading(page.delete);
+            PrimeSelenium.waitGui().until(PrimeExpectedConditions.invisibleAndAnimationComplete(dialog));
         }
 
         assertConfiguration(dialog.getWidgetConfiguration());